+++ task2/schedule.py
@@ -2,13 +2,13 @@ import threading, time, signal
 from datetime import timedelta
 import xkcd as xkcd_class
 
-WAIT_TIME_SECONDS = 10
+WAIT_TIME_SECONDS = 3600
 MAX_COMICS = 2355
 
 class ProgramKilled(Exception):
     pass
 
-def foo():
+def save_xkcd_comics():
     xkcd_object = xkcd_class.XKCD_SERVICE(MAX_COMICS)
     xkcd_object.save_comic()
     print('New image added at: {} '.format(time.ctime()))
@@ -37,7 +37,7 @@ class Job(threading.Thread):
 if __name__ == "__main__":
     signal.signal(signal.SIGTERM, signal_handler)
     signal.signal(signal.SIGINT, signal_handler)
-    job = Job(interval=timedelta(seconds=WAIT_TIME_SECONDS), execute=foo)
+    job = Job(interval=timedelta(seconds=WAIT_TIME_SECONDS), execute=save_xkcd_comics)
     job.start()
 
     while True:
